rename TabstripNewxt bhack to Tabstrip
