+++ vuu-ui/packages/vuu-ui-controls/src/tabstrip/index.ts
@@ -1,4 +1,4 @@
-export * from "./TabstripNext";
+export * from "./Tabstrip";
 export * from "./Tab";
 export * from "./TabsTypes";
 export * from "./tabstrip-dom-utils";
